clean up test changes
