+++ nucliadb/nucliadb/search/tests/fixtures.py
@@ -17,7 +17,6 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-
 import asyncio
 from enum import Enum
 from os.path import dirname
@@ -25,13 +24,17 @@ from typing import Dict, List, Optional
 
 import pytest
 from httpx import AsyncClient
+from nucliadb_protos.nodereader_pb2 import GetShardRequest
+from nucliadb_protos.noderesources_pb2 import Shard
 from redis import asyncio as aioredis
 from starlette.routing import Mount
 
 from nucliadb.ingest.cache import clear_ingest_cache
+from nucliadb.ingest.orm import NODES
+from nucliadb.ingest.orm.node import Node
 from nucliadb.ingest.tests.fixtures import broker_resource
+from nucliadb.ingest.utils import get_driver
 from nucliadb.search import API_PREFIX
-from nucliadb.search.tests.utils import inject_message, wait_for_shard
 from nucliadb_utils.utilities import clear_global_cache
 
 
@@ -197,4 +200,46 @@ async def multiple_search_resource(
         await processor.process(message=message, seqid=count)
 
     await wait_for_shard(knowledgebox_ingest, 100)
+    return knowledgebox_ingest
+
+
+async def inject_message(
+    processor, knowledgebox_ingest, message, count: int = 1
+) -> str:
+    await processor.process(message=message, seqid=count)
+    await wait_for_shard(knowledgebox_ingest, count)
+    return knowledgebox_ingest
+
+
+async def wait_for_shard(knowledgebox_ingest: str, count: int) -> str:
+    # Make sure is indexed
+    driver = await get_driver()
+    txn = await driver.begin()
+    shard = await Node.get_current_active_shard(txn, knowledgebox_ingest)
+    if shard is None:
+        raise Exception("Could not find shard")
+    await txn.abort()
+
+    checks: Dict[str, bool] = {}
+    for replica in shard.shard.replicas:
+        if replica.shard.id not in checks:
+            checks[replica.shard.id] = False
+
+    for i in range(30):
+        for replica in shard.shard.replicas:
+            node_obj = NODES.get(replica.node)
+            if node_obj is not None:
+                req = GetShardRequest()
+                req.shard_id.id = replica.shard.id
+                count_shard: Shard = await node_obj.reader.GetShard(req)  # type: ignore
+                if count_shard.resources >= count:
+                    checks[replica.shard.id] = True
+                else:
+                    checks[replica.shard.id] = False
+
+        if all(checks.values()):
+            break
+        await asyncio.sleep(1)
+
+    assert all(checks.values())
     return knowledgebox_ingest
